Modularized CSES, all contributor effort now in Community_Eval_Methods folder
